Partial implementation of the ResourceDraftBuilderGenerator with factory methods
