+++ commercetools-internal-processors/src/main/java/io/sphere/sdk/annotations/processors/generators/AbstractBuilderGenerator.java
@@ -4,7 +4,6 @@ import com.squareup.javapoet.ClassName;
 import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
 import com.squareup.javapoet.TypeName;
-import io.sphere.sdk.annotations.FactoryMethod;
 import io.sphere.sdk.annotations.processors.models.PropertyGenModel;
 import io.sphere.sdk.models.Reference;
 import io.sphere.sdk.models.Referenceable;
@@ -15,10 +14,10 @@ import javax.lang.model.element.Modifier;
 import javax.lang.model.element.TypeElement;
 import javax.lang.model.util.Elements;
 import java.lang.annotation.Annotation;
-import java.util.*;
-import java.util.function.Function;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  * Abstract base class for builder generators.
@@ -45,57 +44,9 @@ abstract class AbstractBuilderGenerator<A extends Annotation> extends AbstractGe
      * Adds the return statement for the builder method.
      *
      * @param builder the builder to add the return statement
-     *
      * @return the builder with the added return statement
      */
     protected abstract MethodSpec.Builder addBuilderMethodReturn(final TypeElement builderType, final MethodSpec.Builder builder);
-
-
-    protected List<MethodSpec> createFactoryMethods(final FactoryMethod[] factoryMethods, final List<PropertyGenModel> properties, final ClassName returnType) {
-        return Stream.of(factoryMethods)
-                .map(f -> createFactoryMethod(f, properties, returnType))
-                .collect(Collectors.toList());
-    }
-
-    protected MethodSpec createFactoryMethod(final FactoryMethod factoryMethod, final List<PropertyGenModel> properties, final ClassName returnType) {
-        final Set<String> factoryParameterNames = Stream.of(factoryMethod.parameterNames()).collect(Collectors.toCollection(LinkedHashSet::new));
-        final Map<String, PropertyGenModel> getterMethodByPropertyName = properties.stream()
-                .collect(Collectors.toMap(PropertyGenModel::getName, Function.identity()));
-        final List<PropertyGenModel> parameterTemplates = factoryParameterNames.stream()
-                .map(getterMethodByPropertyName::get)
-                .collect(Collectors.toList());
-        assert factoryParameterNames.size() == parameterTemplates.size();
-
-        final String callArguments = properties.stream()
-                .map(p -> factoryParameterNames.contains(p.getName()) ? p.getJavaIdentifier() : "null")
-                .collect(Collectors.joining(", "));
-
-        final MethodSpec.Builder builder = MethodSpec.methodBuilder(factoryMethod.methodName()).addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                .returns(returnType)
-                .addJavadoc("Creates a builder initialized with the given values.\n\n");
-        parameterTemplates.forEach(p -> builder.addJavadoc("@param $L initial value for the $L property\n",
-                p.getJavaIdentifier(), p.getJavadocLinkTag()));
-        builder.addJavadoc("@return new builder initialized with the given values\n");
-        return builder
-                .addParameters(createParameters(parameterTemplates, factoryMethod.useLowercaseBooleans(), false))
-                .addCode("return new $L($L);\n", returnType.simpleName(), callArguments)
-                .build();
-    }
-
-    protected MethodSpec createCopyFactoryMethod(final TypeElement typeElement, final ClassName returnType, final List<ExecutableElement> propertyMethods) {
-        final ParameterSpec templateParameter = ParameterSpec.builder(ClassName.get(typeElement), "template", Modifier.FINAL).build();
-        final String callArguments = propertyMethods.stream()
-                .map(getterMethod -> String.format("template.%s()", getterMethod.getSimpleName()))
-                .collect(Collectors.joining(", "));
-        return MethodSpec.methodBuilder("of").addModifiers(Modifier.PUBLIC, Modifier.STATIC)
-                .returns(returnType)
-                .addJavadoc("Creates a builder initialized with the fields of the template parameter.\n\n")
-                .addJavadoc("@param template the template\n")
-                .addJavadoc("@return a new builder initialized from the template\n")
-                .addParameter(templateParameter)
-                .addCode("return new $L($L);\n", returnType.simpleName(), callArguments)
-                .build();
-    }
 
     protected MethodSpec createBuildMethod(final TypeName returnType, final TypeName implType, final List<ExecutableElement> propertyMethods) {
         final List<String> argumentNames = propertyMethods.stream()
