+++ commercetools-internal-processors/src/test/resources/io/sphere/sdk/annotations/processors/generators/examples/CategoryDraftDsl.java.expected
@@ -1,0 +1,123 @@
+@javax.annotation.Generated(
+    value = "io.sphere.sdk.annotations.processors.generators.ResourceDraftValueGenerator",
+    comments = "Generated from: io.sphere.sdk.annotations.processors.generators.examples.ExampleDraft"
+)
+public final class ExampleDraftDsl extends io.sphere.sdk.models.Base implements io.sphere.sdk.annotations.processors.generators.examples.ExampleDraft {
+  private java.lang.Boolean _default;
+
+  private java.lang.Boolean active;
+
+  private java.util.List<java.lang.String> names;
+
+  @javax.annotation.Nullable
+  private java.lang.String optional;
+
+  private java.lang.String sku;
+
+  @com.fasterxml.jackson.annotation.JsonCreator
+  ExampleDraftDsl(final java.lang.Boolean _default, final java.lang.Boolean active,
+      final java.util.List<java.lang.String> names,
+      @javax.annotation.Nullable final java.lang.String optional, final java.lang.String sku) {
+    this._default = _default;
+    this.active = active;
+    this.names = names;
+    this.optional = optional;
+    this.sku = sku;
+  }
+
+  public java.lang.Boolean isDefault() {
+    return _default;
+  }
+
+  public java.lang.Boolean isActive() {
+    return active;
+  }
+
+  public java.util.List<java.lang.String> getNames() {
+    return names;
+  }
+
+  @javax.annotation.Nullable
+  public java.lang.String getOptional() {
+    return optional;
+  }
+
+  public java.lang.String getSku() {
+    return sku;
+  }
+
+  /**
+   * Creates a new instance of {@code io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftBuilder} with the values of this builder.
+   *
+   * @return the instance
+   */
+  public io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftBuilder newBuilder() {
+    return new io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftBuilder(_default, active, names, optional, sku);
+  }
+
+  public io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl with_default(final java.lang.Boolean _default) {
+    return newBuilder()._default(_default).build();
+  }
+
+  public io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl withActive(final java.lang.Boolean active) {
+    return newBuilder().active(active).build();
+  }
+
+  public io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl withNames(final java.util.List<java.lang.String> names) {
+    return newBuilder().names(names).build();
+  }
+
+  public io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl withOptional(@javax.annotation.Nullable final java.lang.String optional) {
+    return newBuilder().optional(optional).build();
+  }
+
+  public io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl withSku(final java.lang.String sku) {
+    return newBuilder().sku(sku).build();
+  }
+
+  /**
+   * Creates a builder initialized with the given values.
+   *
+   * @param sku initial value for the {@link ExampleDraft#getSku()} property
+   * @param active initial value for the {@link ExampleDraft#isActive()} property
+   * @return new builder initialized with the given values
+   */
+  public static io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl of(final java.lang.String sku,
+      final java.lang.Boolean active) {
+    return new ExampleDraftDsl(null, active, null, null, sku);
+  }
+
+  /**
+   * Creates a builder initialized with the given values.
+   *
+   * @param _default initial value for the {@link ExampleDraft#isDefault()} property
+   * @param active initial value for the {@link ExampleDraft#isActive()} property
+   * @return new builder initialized with the given values
+   */
+  public static io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl of(final java.lang.Boolean _default,
+      final java.lang.Boolean active) {
+    return new ExampleDraftDsl(_default, active, null, null, null);
+  }
+
+  /**
+   * Creates a builder initialized with the given values.
+   *
+   * @param _default initial value for the {@link ExampleDraft#isDefault()} property
+   * @param optional initial value for the {@link ExampleDraft#getOptional()} property
+   * @return new builder initialized with the given values
+   */
+  public static io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl of(final java.lang.Boolean _default,
+      final java.lang.String optional) {
+    return new ExampleDraftDsl(_default, null, null, optional, null);
+  }
+
+  /**
+   * Creates a builder initialized with the fields of the template parameter.
+   *
+   * @param template the template
+   * @return a new builder initialized from the template
+   */
+  public static io.sphere.sdk.annotations.processors.generators.examples.ExampleDraftDsl of(final io.sphere.sdk.annotations.processors.generators.examples.ExampleDraft template) {
+    return new ExampleDraftDsl(template.isDefault(), template.isActive(), template.getNames(), template.getOptional(), template.getSku());
+  }
+}
